[Step3] refactor : 리뷰 반영 - 테스트 코드 패키지 경로, 클래스명 수정
