+++ src/test/kotlin/lotto/agency/LottoTicketTest.kt
@@ -1,7 +1,5 @@
-package lotto
+package lotto.agency
 
-import lotto.agency.LottoNumber
-import lotto.agency.LottoTicket
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
